Only Perform Configuration Check if Project is Defined
The 'xcodebuild -list' command will not list out configurations for
workspaces, which causes the configuration validation to fail. With
this change, if you're building a workspace, xcodebuild will use the
default configuration (seems to be Release, which should be fine).
However, you can still specify Debug in the ipa options and it will
adhere to them.