Add support for building Swift apps
The .ipa of a Swift application must contain a "SwiftSupport" directory
with all the runtime libraries required by Swift (even though they are
already included in the application bundle's "Frameworks" directory).

This step is NOT done by Xcode. The "Organizer" does that when someone
exports the application for distribution to iTunes Connect.

By inspecting the .ipa produced by the Organizer, we see that the SHA1
of all Swift runtime libraries in the "SwiftSupport" directory
are different from the ones inside the app bundle's "Frameworks"
directory. They seem to come from Xcode's toolchain directory instead.

This commit tries to match the behavior of the Organizer by doing the
following (after the PackageApplication step):

- Inspect the "Frameworks" directory within the app bundle to determine
  which runtime libraries we need (if any)
- Copy all Swift runtime libraries from Xcode's toolchain directory in
  a temporary directory with the required directory structure.
- Append the temporary "SwiftSupport" directory to the .ipa file