Path to generated ipa file is last output
The absolute path is the last line displayed, which makes it easier for other tools (thinking about [fastlane](http://fastlane.tools) to parse the path to the generated ipa file. 

If anybody finds a better way to pass the path to the generated ipa (e.g. using environment variables), let me know.