Fixed options "clean, build, archive" evaluation
"clean, build, archive" were only checked for false. If a argument is not specified the argument is nil, so the check for false will fail. The "build" argument will only be added if "archive" is not specified, because if not, xcodebuild will build the app two times (one time for "build" and one time for "archive").